as(): make as() cast to use type_safety.enforce
The current implementation for std::any, std::optional, and std::variant
throws exceptions. This behaviour is inconsistent to runtime checked
as() for signed/unsigned integral cast.

This implementation allows to setup one approch for type_safety issues.
Current behaviour is to terminate but it can be changed to e.g. throwing exception.