Removed ban on using the name of a type scope name elsewhere in a type
See #1373

The programmer can always write `this.` to disambiguate. This originally seemed like a useful constraint, but it creates a tension between generated code and user code. For example, see #1373, and the example

    foo : @enum type = { s; }

which before this commit was rejected because the `@enum` type metafunction generates a function with a parameter named `s`.

It seems undesirable to force generated code to uglify its names to avoid conflicts.

Also, fix `@enum` for cases like this to type-scope-qualify access to the member.