Add `is` and `as` support for `std::expected` (v2) (#971)
* Add `is` and `as` support for `std::expected`

* Add test-results for `pure2-expected-is-as` test case

* Update test-results

* [CI] run-tests.sh now allows each compiler config to exclude test files that it doesn't want to run

There are 2 macOS test configs which share the same expected results directory (to avoid duplication). This works great except for the new `pure2-expected-is-as` test. This new test code fails to compile (as expected) on both compilers, but produces a slightly different error diagnostic because the path is different.

E.g.

/Applications/Xcode_14.3.1.app/...etc.../math.h
and
/Library/Developer/CommandLineTools/...etc.../math.h

One option would be to stop sharing the expected results for both of these compilers, but that seems wasteful since it's just one test which fails to compile. So instead the `run-tests` script has a new way to exclude a test from running.